create, update and delete posts

+++ app/Http/Controllers/PostsController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Exception;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 class PostsController extends Controller
 {
@@ -25,7 +27,28 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-        return Post::create($request->all());
+        try {
+            $this->validate($request, [
+                'title' => 'required|max:200',
+                'content' => 'required|max:4000',
+                'category_id' => 'required',
+            ]);
+        } catch (Exception $e) {
+            return response(['message' => 'Invalid format of data'], 400);
+        }
+
+        $user = JWTAuth::user();
+
+        if (!$user) {
+            return response(['message' => 'Token required'], 400);
+        }
+
+        Post::create(([
+            'user_id' => $user->id,
+            'title' => $request['title'],
+            'content' => $request['content'],
+            'category_id' => $request['category_id']
+        ]));
     }
 
     /**
@@ -52,9 +75,22 @@ class PostsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $user = JWTAuth::user();
+        
+        if (!$user) {
+            return response(['message' => 'Token required'], 400);
+        }
+
         $post = Post::find($id);
-        $post->update($request->all());
-        return $post;
+        
+        if ($post == null) {
+            return response(['message' => 'No such post'], 404);
+        } else if ($user['role'] != 'admin' && $post['user_id'] != $user['id']) {
+            return response(['message' => 'Can\'t update others info'], 403);
+        } else {
+            $post->update($request->all());
+            return $post;
+        }
     }
 
     /**
@@ -65,6 +101,20 @@ class PostsController extends Controller
      */
     public function destroy($id)
     {
-        return Post::destroy($id);
+        $user = JWTAuth::user();
+        
+        if (!$user) {
+            return response(['message' => 'Token required'], 400);
+        }
+
+        $post = Post::find($id);
+        
+        if ($post == null) {
+            return response(['message' => 'No such post'], 404);
+        } else if ($user['role'] != 'admin' && $post['user_id'] != $user['id']) {
+            return response(['message' => 'Can\'t delete others info'], 403);
+        } else {
+            return Post::destroy($id);
+        }
     }
 }
